Move CapabilityChecker to covia.lattice
Capability checking is a lattice/UCAN concern, not an adapter concern —
the resource and ability vocabulary it operates on belongs to the lattice
addressing model. Moving it out of covia.adapter clears the way for
adapters to depend on it without circling back into the adapter package
namespace, and groups it with the other lattice helpers.

Tidy LLMAgentAdapter while in the file: drops two now-stale imports
(CapabilityChecker and ContextInspectable, which moved alongside this).
